+++ fastx_barber/scripts/__init__.py
@@ -4,12 +4,14 @@
 """
 
 from fastx_barber.scripts import arguments
-from fastx_barber.scripts import extract, match
+from fastx_barber.scripts import extract, flag_filter
+from fastx_barber.scripts import match
 from fastx_barber.scripts import trim, trim_len, trim_qual, trim_regex
 
 __all__ = [
     "arguments",
     "extract",
+    "flag_filter",
     "match",
     "trim",
     "trim_len",