+++ fastx_barber/scripts/trim_len.py
@@ -81,7 +81,7 @@ def parse_arguments(args: argparse.Namespace) -> argparse.Namespace:
     args = scriptio.set_tempdir(args)
 
     if 0 == args.length:
-        logging.info("Trimming length equal to 0. Nothing to do.")
+        logging.info("Trimming length equal to 0. Nothing to do. :person_shrugging:")
         sys.exit()
 
     if args.log_file is not None: